pyproject.toml
[project]
name = "stocks-analysis"
version = "0.1.0"
description = "Fetch Zerodha Kite holdings, store in Google Sheets, analyze portfolio"
requires-python = ">=3.13"
dependencies = [
    "playwright",
    "gspread",
    "pandas",
    "scipy",
    "matplotlib",
    "python-dotenv",
    "yfinance",
    "numpy",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-cov",
    "ruff",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = ["integration: marks tests requiring external services"]
addopts = "-v"

[tool.coverage.run]
source = ["stocks_analysis"]

[tool.coverage.report]
fail_under = 80
show_missing = true

[tool.ruff]
target-version = "py313"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.ruff.format]
quote-style = "double"
